config.py renamed to ExportItFrameworkConfig.py to avoid conflicts with other add-ins that are using the https://github.com/tapnair/apper framework. This fixes #14
